auth1/serializers.py
from rest_framework import serializers
from auth1.models import User, Student, Course

class UserSerializer(serializers.ModelSerializer):
    class Meta:
        model = User
        fields = ['id','username', 'email', 'role', 'password']
        extra_kwargs = {'password': {'write_only': True}}

    def create(self, validated_data):
        user = User.objects.create_user(**validated_data)
        return user

class StudentSerializer(serializers.ModelSerializer):
    class Meta:
        model = Student
        fields = [
            'user', 
            'date_of_birth', 
            'enrollment_date', 
            'is_active', 
            'student_id', 
            'grade'
        ]
class CourseSerializer(serializers.ModelSerializer):
    class Meta:
        model = Course
        fields = '__all__'